Move Apply components into their own directory
